power-utils: cancel suspend if sleep delay inhibitor locks are held
There may be cases where a suspend operation is requested, but a
userspace service is not able to successfully prepare for suspend. In
this case, with this change, the suspend operation is cancelled.

This allows us to fail the suspend operation fast in userspace, rather
than triggering suspend and causing confusing / hard-to-debug behavior
downstream in the suspend operation.

See power-utils/README.txt for implementation details

Change-Id: I19cdd93809307e6a79811ee7ccabaab45e350f43
Signed-off-by: Eric Smith <quic_ericsmit@quicinc.com>